backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
pydantic-settings==2.8.1
python-dotenv==1.0.1
httpx==0.28.1
openai==1.64.0
praw==7.8.1
apscheduler==3.10.4
scikit-learn==1.6.1
numpy==2.2.3
python-jose[cryptography]==3.3.0
email-validator==2.2.0
orjson==3.10.15
